+++ unstructured/partition/auto.py
@@ -249,16 +249,11 @@ def partition(
         )
         return augment_metadata(elements)
 
-    # -- JSON is a special case because it's not a document format per se and is insensitive to
-    # -- most of the parameters that apply to other file types.
-    if file_type == FileType.JSON:
-        partition_json = partitioner_loader.get(file_type)
-        elements = partition_json(filename=filename, file=file, **kwargs)
-        return augment_metadata(elements)
-
-    if file_type == FileType.NDJSON:
-        partition_ndjson = partitioner_loader.get(file_type)
-        elements = partition_ndjson(filename=filename, file=file, **kwargs)
+    # -- JSON/NDJSON are special cases: not document formats per se and insensitive to most
+    # -- parameters that apply to other file types.
+    if file_type in (FileType.JSON, FileType.NDJSON):
+        partitioner = partitioner_loader.get(file_type)
+        elements = partitioner(filename=filename, file=file, **kwargs)
         return augment_metadata(elements)
 
     # -- EMPTY is also a special case because while we can't determine the file type, we can be